+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## [Unreleased]
+## [1.2.0] 2023-12-21
 ### Added
+- booking id for delivery
 - Example for retrieval
 - URL split for url usage
 - Destroy web view when finish releasing memory and data